Share the merge base validation, error and Bitbucket request helpers
