+++ vcsclient/bitbucketcloud.go
@@ -1253,12 +1253,7 @@ func (client *BitbucketCloudClient) downloadRepositoryViaGitClone(ctx context.Co
 
 // GetMergeBase on Bitbucket cloud
 func (client *BitbucketCloudClient) GetMergeBase(ctx context.Context, owner, repository, refBefore, refAfter string) (commitInfo CommitInfo, err error) {
-	if err = validateParametersNotBlank(map[string]string{
-		"owner":      owner,
-		"repository": repository,
-		"refBefore":  refBefore,
-		"refAfter":   refAfter,
-	}); err != nil {
+	if err = validateMergeBaseParameters(owner, repository, refBefore, refAfter); err != nil {
 		return
 	}
 
@@ -1273,27 +1268,9 @@ func (client *BitbucketCloudClient) GetMergeBase(ctx context.Context, owner, rep
 	if err != nil {
 		return
 	}
-	request, err := http.NewRequestWithContext(ctx, http.MethodGet, requestUrl, nil)
+	body, err := getBitbucketJson(ctx, bitbucketClient.HttpClient, requestUrl, client.setAuthenticationHeader)
 	if err != nil {
 		return
-	}
-	client.setAuthenticationHeader(request)
-
-	response, err := bitbucketClient.HttpClient.Do(request)
-	if err != nil {
-		return
-	}
-	defer func() {
-		err = errors.Join(err, response.Body.Close())
-	}()
-
-	body, err := io.ReadAll(response.Body)
-	if err != nil {
-		return
-	}
-	if response.StatusCode != http.StatusOK {
-		return CommitInfo{}, fmt.Errorf("failed to get the merge base of '%s' and '%s' in <%s/%s>, status: %s, body: %s",
-			refBefore, refAfter, owner, repository, response.Status, body)
 	}
 
 	var mergeBase commitDetails
@@ -1301,7 +1278,7 @@ func (client *BitbucketCloudClient) GetMergeBase(ctx context.Context, owner, rep
 		return
 	}
 	if mergeBase.Hash == "" {
-		return CommitInfo{}, fmt.Errorf("no merge base found for <%s/%s> between %s and %s", owner, repository, refBefore, refAfter)
+		return CommitInfo{}, mergeBaseNotFoundError(owner, repository, refBefore, refAfter)
 	}
 	return mapBitbucketCloudCommitToCommitInfo(mergeBase), nil
 }
